feat: Delete note operation added

+++ app.js
@@ -3,21 +3,42 @@ const yargs = require('yargs');
 const chalk = require('chalk');
 
 const command = process.argv[2];
-const note_title = yargs.argv['title'];
-const note_content = yargs.argv['body'];
-const note_file_name = yargs.argv['file'];
 
 // \
 
+const note_file_name = 'notes/' + yargs.argv['file'];
+
 addNote = () => {
+	const fileExists = fs.existsSync(note_file_name);
+	if (fileExists) {
+		logDanger('Error: file already exists...');
+		return;
+	}
+	const note_title = yargs.argv['title'];
+	const note_content = yargs.argv['body'];
+	
 	fs.writeFileSync(note_file_name, note_title.toUpperCase());
 	fs.appendFileSync(note_file_name, "\n");
 	fs.appendFileSync(note_file_name, note_content);
-	console.log(chalk.inverse.bold.green(' New note successfully created!!! '));
+	logSuccess('New note successfully created!!!');
 }
 
 deleteNote = () => {
-	console.log('Deleting note...');
+	try {
+		logDanger('Deleting note: ' + note_file_name);
+		fs.rmSync(note_file_name);
+		logSuccess('Note successfully deleted!!!');
+	} catch(error) {
+		logDanger('Error: ' + error.message);
+	}
+}
+
+logSuccess = (message) => {
+	console.log(chalk.green.bold.inverse(' ' + message + ' '));
+}
+
+logDanger = (message) => {
+	console.log(chalk.red.bold.inverse(' ' + message + ' '));
 }
 
 switch(command){
@@ -28,5 +49,5 @@ switch(command){
 		deleteNote();
 		break;
 	default:
-		console.log('default option...');
+		logDanger('Invalid option. Try agin...');
 }
